# Conflicts:
#	.vs/Fubon.Treasury/v14/.suo
#	Treasury.Web/Enum/OpenPartialViewType.cs
#	Treasury.Web/Treasury.Web.csproj
#	Treasury.Web/bin/Treasury.Web.dll
#	Treasury.Web/bin/Treasury.Web.pdb
#	Treasury.Web/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
#	Treasury.Web/obj/Debug/RdlCompile.cache
#	Treasury.Web/obj/Debug/TempPE/Models.dbTreasury.Designer.cs.dll
#	Treasury.Web/obj/Debug/TempPE/Models.dbTreasury.cs.dll
#	Treasury.Web/obj/Debug/Treasury.Web.csprojResolveAssemblyReference.cache
#	Treasury.Web/obj/Debug/Treasury.Web.dll
#	Treasury.Web/obj/Debug/Treasury.Web.pdb

+++ Treasury.Web/Service/Actual/Marging.cs
@@ -11,6 +11,7 @@ using Treasury.WebDaos;
 using Treasury.WebUtility;
 using Treasury.Web.Enum;
 using System.ComponentModel;
+using Treasury.Web.Controllers;
 
 /// <summary>
 /// 功能說明：金庫進出管理作業-金庫物品存取申請作業 存出保證金
@@ -144,6 +145,9 @@ namespace Treasury.Web.Service.Actual
                     result.AddRange(db.ITEM_REFUNDABLE_DEP.AsNoTracking()
                         .Where(x => TreasuryIn.Contains(x.INVENTORY_STATUS), searchModel.vTreasuryIO == "Y")
                         .Where(x => x.INVENTORY_STATUS == TreasuryOut, searchModel.vTreasuryIO == "N")
+                        .Where(x => x.ITEM_ID == searchModel.vItem_No, !searchModel.vItem_No.IsNullOrWhiteSpace())
+                        .Where(x => x.BOOK_NO == searchModel.vBookNo, !searchModel.vBookNo.IsNullOrWhiteSpace())
+                        .Where(x => x.MARGIN_DEP_TYPE == searchModel.vMargin_Dep_Type, searchModel.vMargin_Dep_Type != "All")
                         .Where(x => x.PUT_DATE != null && x.PUT_DATE.Value >= PUT_DATE_From.Value, PUT_DATE_From != null)
                         .Where(x => x.PUT_DATE != null && x.PUT_DATE.Value <= PUT_DATE_To.Value, PUT_DATE_To != null)
                         .Where(x => x.GET_DATE != null && x.GET_DATE.Value >= GET_DATE_From.Value, GET_DATE_From != null)
@@ -153,8 +157,10 @@ namespace Treasury.Web.Service.Actual
                         {
                             vItem_PK = x.ITEM_ID,
                             vStatus = x.INVENTORY_STATUS,
-                            vPut_Date = x.PUT_DATE?.ToString("yyyy/MM/dd"),
+                            vPut_Date = x.PUT_DATE?.dateTimeToStr(),
+                            vGet_Date = x.GET_DATE?.dateTimeToStr(),
                             vBook_No = x.BOOK_NO,
+                            vBook_No_AFT = x.BOOK_NO_AFT,
                             vAply_Uid = x.APLY_UID,
                             vAply_Uid_Name = emps.FirstOrDefault(y => y.USR_ID == x.APLY_UID)?.EMP_NAME?.Trim(),
                             vCHARGE_DEPT = x.CHARGE_DEPT,
@@ -176,6 +182,18 @@ namespace Treasury.Web.Service.Actual
                             vMemo_AFT = x.MEMO_AFT,
                             vLast_Update_Time = x.LAST_UPDATE_DT
                         }).ToList());
+                    if (searchModel.vTreasuryIO == "N") //取出
+                    {
+                        if (result.Any())
+                        {
+                            var itemIds = result.Select(x => x.vlItem_Id).ToList();
+                            var uids = GetAplyUidName(itemIds);
+                            result.ForEach(x =>
+                            {
+                                x.vGet_Uid_Name = uids.FirstOrDefault(y => y.itemId == x.vlItem_Id)?.getAplyUidName;
+                            });
+                        }
+                    }
                 }
                 else
                 {
@@ -188,8 +206,10 @@ namespace Treasury.Web.Service.Actual
                         {
                             vItem_PK = x.ITEM_ID,
                             vStatus = x.INVENTORY_STATUS,
-                            vPut_Date = x.PUT_DATE?.ToString("yyyy/MM/dd"),
+                            vPut_Date = x.PUT_DATE?.dateTimeToStr(),
+                            vGet_Date = x.GET_DATE?.dateTimeToStr(),
                             vBook_No = x.BOOK_NO,
+                            vBook_No_AFT = x.BOOK_NO_AFT,
                             vAply_Uid = x.APLY_UID,
                             vAply_Uid_Name = emps.FirstOrDefault(y => y.USR_ID == x.APLY_UID)?.EMP_NAME?.Trim(),
                             vCHARGE_DEPT = x.CHARGE_DEPT,
@@ -389,8 +409,17 @@ namespace Treasury.Web.Service.Actual
                             {
                                 #region 申請單紀錄檔
                                 _TAR = db.TREA_APLY_REC.First(x => x.APLY_NO == taData.vAplyNo);
-                                if (_TAR.APLY_STATUS != _APLY_STATUS) //申請紀錄檔狀態不是在表單申請狀態
-                                    _APLY_STATUS = Ref.AccessProjectFormStatus.A05.ToString(); //為重新申請案例
+                                if (CustodyAppr.Contains(_TAR.APLY_STATUS))
+                                {
+                                    _APLY_STATUS = CustodyConfirmStatus;
+                                    _TAR.CUSTODY_UID = AccountController.CurrentUserId; //保管單位直接帶使用者
+                                    _TAR.CUSTODY_DT = dt;
+                                }
+                                else
+                                {
+                                    if (_TAR.APLY_STATUS != _APLY_STATUS) //申請紀錄檔狀態不是在表單申請狀態
+                                        _APLY_STATUS = Ref.AccessProjectFormStatus.A05.ToString(); //為重新申請案例
+                                }
                                 _TAR.APLY_STATUS = _APLY_STATUS;
                                 _TAR.LAST_UPDATE_DT = dt;
 
@@ -402,7 +431,7 @@ namespace Treasury.Web.Service.Actual
                                 new APLY_REC_HIS()
                                 {
                                     APLY_NO = _TAR.APLY_NO,
-                                    APLY_STATUS = _TAR.APLY_STATUS,
+                                    APLY_STATUS = _APLY_STATUS,
                                     PROC_UID = _TAR.CREATE_UID,
                                     PROC_DT = dt
                                 });
@@ -511,7 +540,7 @@ namespace Treasury.Web.Service.Actual
                                             }
                                         }
                                     }
-                                    else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString()) //判斷申請作業-取出
+                                    else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString() && (_APLY_STATUS != CustodyConfirmStatus)) //判斷申請作業-取出
                                     {
                                         _IRD = db.ITEM_REFUNDABLE_DEP.FirstOrDefault(x => x.ITEM_ID == item.vItem_Id);
                                         if (_IRD.LAST_UPDATE_DT > item.vLast_Update_Time)
@@ -558,7 +587,7 @@ namespace Treasury.Web.Service.Actual
                                     }));
                                     db.ITEM_REFUNDABLE_DEP.AddRange(inserts);
                                 }
-                                else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString())//取出
+                                else if (taData.vAccessType == Ref.AccessProjectTradeType.G.ToString() && (_APLY_STATUS != CustodyConfirmStatus))//取出
                                 {
                                     db.OTHER_ITEM_APLY.RemoveRange(db.OTHER_ITEM_APLY.Where(x => x.APLY_NO == taData.vAplyNo).ToList());
                                     db.OTHER_ITEM_APLY.AddRange(updateItemIds.Select(x => new OTHER_ITEM_APLY()
@@ -712,6 +741,7 @@ namespace Treasury.Web.Service.Actual
                             _Marging.DESCRIPTION_AFT = model.vDescription_AFT;
                             _Marging.MEMO_AFT = model.vMemo_AFT;
                             _Marging.LAST_UPDATE_DT = dt;
+                            _Marging.BOOK_NO_AFT = model.vBook_No_AFT;
 
                             logStr = _Marging.modelToString(logStr);
 
@@ -770,6 +800,7 @@ namespace Treasury.Web.Service.Actual
                 if (_Marging != null)
                 {
                     _Marging.INVENTORY_STATUS = "1"; //在庫
+                    _Marging.BOOK_NO_AFT = null;
                     _Marging.MARGIN_DEP_TYPE_AFT = null;
                     _Marging.TRAD_PARTNERS_AFT = null;
                     _Marging.AMOUNT_AFT = null;
@@ -815,6 +846,8 @@ namespace Treasury.Web.Service.Actual
                     _Marging.DESCRIPTION_AFT = null;
                     _Marging.MEMO = GetNewValue(_Marging.MEMO, _Marging.MEMO_AFT);
                     _Marging.MEMO_AFT = null;
+                    _Marging.BOOK_NO = GetNewValue(_Marging.BOOK_NO, _Marging.BOOK_NO_AFT);
+                    _Marging.BOOK_NO_AFT = null;
                     _Marging.LAST_UPDATE_DT = dt;
                     logStr = _Marging.modelToString(logStr);
                 }
